requirements.txt
torch==1.11.0
torchvision==0.12.0
torchaudio==0.11.0

timm
numpy
scipy
matplotlib
Pillow
PyYAML
onnxruntime
